+++ cpp/src/operators/c2r/ColumnarToRow.h
